# TRegex: Add ASCII encoding to reduce range of Unicode char sets to ASCII if match string is ASCII-only.