--- a/regex/CHANGELOG.md
+++ b/regex/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This changelog summarizes major changes between TRegex versions relevant to language implementors integrating TRegex into their language. This document will focus on API changes relevant to integrators of TRegex.
 
+## Version 22.0.0
+
+* Added new `ASCII` encoding that callers can use when compiling a regex to limit the range of code point matches to [0xx, 0x7f].
+
 ## Version 21.3.0
 
 * Support for case-insensitive matching in Ruby regular expressions.