--- a/regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/string/Encodings.java
+++ b/regex/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/string/Encodings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -62,8 +62,9 @@ public final class Encodings {
     public static final Encoding UTF_32 = new Encoding.UTF32();
     public static final Encoding UTF_16_RAW = new Encoding.UTF16Raw();
     public static final Encoding LATIN_1 = new Encoding.Latin1();
+    public static final Encoding ASCII = new Encoding.Ascii();
 
-    public static final String[] ALL_NAMES = {UTF_8.getName(), UTF_16.getName(), UTF_16_RAW.getName(), UTF_32.getName(), LATIN_1.getName(), "BYTES"};
+    public static final String[] ALL_NAMES = {UTF_8.getName(), UTF_16.getName(), UTF_16_RAW.getName(), UTF_32.getName(), ASCII.getName(), LATIN_1.getName(), "BYTES"};
 
     public static Encoding getEncoding(String name) {
         switch (name) {
@@ -496,6 +497,64 @@ public final class Encodings {
             @Override
             public StringBufferLATIN1 createStringBuffer(int capacity) {
                 return new StringBufferLATIN1(capacity);
+            }
+
+            @Override
+            public LoopOptimizationNode extractLoopOptNode(CodePointSet cps) {
+                return new LoopOptIndexOfAnyByteNode(cps.inverseToByteArray(this));
+            }
+
+            @Override
+            public int getNumberOfDecodingSteps() {
+                return 1;
+            }
+
+            @Override
+            public void createMatcher(Builder matchersBuilder, int i, CodePointSet cps, CompilationBuffer compilationBuffer) {
+                matchersBuilder.getBuffer(0).set(i, CharMatchers.createMatcher(cps, compilationBuffer));
+            }
+
+            @Override
+            public Matchers toMatchers(Builder matchersBuilder) {
+                return new Matchers.SimpleMatchers(matchersBuilder.materialize(0), matchersBuilder.getNoMatchSuccessor());
+            }
+        }
+
+        public static final class Ascii extends Encoding {
+
+            @Override
+            public String getName() {
+                return "ASCII";
+            }
+
+            @Override
+            public int getMaxValue() {
+                return 0x7f;
+            }
+
+            @Override
+            public CodePointSet getFullSet() {
+                return Constants.ASCII_RANGE;
+            }
+
+            @Override
+            public int getEncodedSize(int codepoint) {
+                return 1;
+            }
+
+            @Override
+            public boolean isFixedCodePointWidth(CodePointSet set) {
+                return true;
+            }
+
+            @Override
+            public boolean isUnicode() {
+                return false;
+            }
+
+            @Override
+            public AbstractStringBuffer createStringBuffer(int capacity) {
+                return new StringBufferASCII(capacity);
             }
 
             @Override