This PR adds a new ASCII encoding to TRegex, which has half the range of valid values as Latin-1. If a guest language knows the regex consists only of ASCII characters, it can specify that when constructing a regex object.

TRegex already has a good mechanism to restrict the code point range for things like POSIX bracket expressions based on the regex's associated encoding. The new ASCII encoding hooks into that existing mechanism to further limit the set of code point ranges based on data the caller knows at regex creation time. Below is a table of the number of code point ranges for the set of POSIX bracket expressions, along with two non-POSIX bracket expressions that the Ruby language uses:

| POSIX Bracket Expr. | Full Unicode # of Ranges | LATIN-1 Range # of Ranges | ASCII # of Ranges|
|-----------------------------|:----------------------------------:|:-------------------------------------:|:------------------------:|
| [:alnum] | 732 | 9 | 3 |
| [:alpha:] | 695 | 8 | 2 |
| [:blank:] | 8 | 3 | 2 |
| [:cntrl:] | 2 | 2 | 2 |
| [:digit:] | 61 | 1 | 1 |
| [:graph:] | 682 | 2 | 1 |
| [:lower:] | 652 | 6 | 1 |
| [:print:] | 679 | 2 | 1 |
| [:punct:] | 334 | 13 | 4 |
| [:space:] | 10 | 4 | 2 |
| [:upper:] | 643 | 3 | 1 |
| [:xdigit:] | 3 | 3 | 3 |
| ↓ Non-POSIX Bracket Expr. ↓ |-----|---|----|
| [:ascii:] | 1 | 1 | 1 |
| [:word:] | 733 | 10 | 4 |

In some cases, such as `[:xdigit:]` there's no gain, but in others such as `[:lower:]` we're able to reduce the entire set to a single range. TRegex doesn't optimize for single range code point sets specifically, but the membership operation has fewer sets to search through.

I benchmarked with TruffleRuby, doing a simple one-or-more check on the bracket expression (e.g., `/[[:alnum:]]+/`). On typical inputs, the savings was not drastic (3 - 6%) for some character classes (`[:alpha:]`, `[:alnum:]`, `[:lower:]`, and `[:space:]`). On problematic inputs, such as a repeating string of `"{}"`, the savings was more pronounced (20% - 128% faster). I did not spend a great deal of time looking for pathological cases. The execution difference is a function of both the regexp pattern and the length and composition of the match string. The code point set membership check performs a binary search to determine if a codepoint belongs to any of its sorted code point ranges. As such, problematic characters could force the search procedure to check the maximum number of ranges. But, at the same time, that execution time is logarithmic.

Overall, I don't expect to see a drastic speed-up on typical workloads. Although, I reserve the right to be incorrect about that as I didn't explore regexp pattern space in great detail. However, I think this is still a worthwhile change because regexp matches are often performed against user-supplied data and we wouldn't want an adversary to force the engine down a bad path.