@jirkamarsik 2021-09-20
Thanks for your contribution and its detailed analysis! This looks great! I'll look for ways how we can merge this. :-)