@djoooooe 2021-09-20
Just to clarify: The binary-search code point set membership check as seen in the class `CodePointSet` is never used in the actual compiled code, this class is just an intermediate representation used by the DFA generator.