@jirkamarsik 2021-09-20
Do you have the benchmarks so that I can try reproducing the results locally?

I used a benchmark like this:
```
require 'benchmark/ips'

input = ("{}" * 100).force_encoding("us-ascii")

Benchmark.ips do |x|
  x.iterations = 2

  ['alnum', 'alpha', 'blank', 'cntrl', 'digit', 'graph', 'lower', 'print', 'punct', 'space', 'upper', 'xdigit', 'ascii', 'word'].each do |char_class_name|
    char_class = "[[:#{char_class_name}:]]"
    re = Regexp.new("#{char_class}+".force_encoding("us-ascii"))
    x.report(char_class) do |times|
      i = 0
      while i < times
        re.match(input)
        i += 1
      end
    end
  end

  x.compare!
end
```
and I got the following results.

Without the `Ascii` encoding:
```
         [[:blank:]]: 39014490.4 i/s
         [[:ascii:]]:  4918644.9 i/s - 7.93x  (± 0.00) slower
         [[:graph:]]:  4179991.4 i/s - 9.33x  (± 0.00) slower
         [[:print:]]:  4127884.4 i/s - 9.45x  (± 0.00) slower
         [[:punct:]]:  3383031.9 i/s - 11.53x  (± 0.00) slower
         [[:digit:]]:  3071249.7 i/s - 12.70x  (± 0.00) slower
         [[:cntrl:]]:  3035635.3 i/s - 12.85x  (± 0.00) slower
          [[:word:]]:  2696047.1 i/s - 14.47x  (± 0.00) slower
         [[:alnum:]]:  2695568.4 i/s - 14.47x  (± 0.00) slower
         [[:upper:]]:  2647868.3 i/s - 14.73x  (± 0.00) slower
         [[:alpha:]]:  2638921.0 i/s - 14.78x  (± 0.00) slower
         [[:lower:]]:  2628058.5 i/s - 14.85x  (± 0.00) slower
        [[:xdigit:]]:  2623925.0 i/s - 14.87x  (± 0.00) slower
         [[:space:]]:  2588324.0 i/s - 15.07x  (± 0.00) slower
```

With the `Ascii` encoding:
```
         [[:blank:]]: 40253751.3 i/s
         [[:ascii:]]:  6025486.0 i/s - 6.68x  (± 0.00) slower
         [[:print:]]:  4897423.4 i/s - 8.22x  (± 0.00) slower
         [[:graph:]]:  4799291.6 i/s - 8.39x  (± 0.00) slower
         [[:alpha:]]:  3667002.0 i/s - 10.98x  (± 0.00) slower
         [[:punct:]]:  3397407.9 i/s - 11.85x  (± 0.00) slower
         [[:lower:]]:  3131768.3 i/s - 12.85x  (± 0.00) slower
         [[:upper:]]:  3127483.9 i/s - 12.87x  (± 0.00) slower
         [[:digit:]]:  2994088.2 i/s - 13.44x  (± 0.00) slower
         [[:space:]]:  2908833.4 i/s - 13.84x  (± 0.00) slower
         [[:cntrl:]]:  2871670.7 i/s - 14.02x  (± 0.00) slower
         [[:alnum:]]:  2810562.2 i/s - 14.32x  (± 0.00) slower
        [[:xdigit:]]:  2678455.9 i/s - 15.03x  (± 0.00) slower
          [[:word:]]:  2633098.2 i/s - 15.29x  (± 0.00) slower
```

This leads to the following observed performance changes:

POSIX Bracket Expr. | Speedup
---------------------------- | -----------
[[:alnum:]] | 1.04x
[[:alpha:]] | 1.39x
[[:blank:]] | 1.03x
[[:cntrl:]] | 0.95x
[[:digit:]] | 0.97x
[[:graph:]] | 1.15x
[[:lower:]] | 1.19x
[[:print:]] | 1.19x
[[:punct:]] | 1.00x
[[:space:]] | 1.12x
[[:upper:]] | 1.18x
[[:xdigit:]] | 1.02x
[[:ascii:]] | 1.22x
[[:word:]] | 0.98x