@nirvdrum 2021-09-20
> Just to clarify: The binary-search code point set membership check as seen in the class `CodePointSet` is never used in the actual compiled code, this class is just an intermediate representation used by the DFA generator.

Thanks for clarifying. I've been trying to gain a better understanding of how all of TRegex works and that's mostly been accomplished by stepping through with a debugger. It's a bit tangential to the issue, but if you could please point me at the relevant specializations, that'd be helpful.